requirements.txt
openai-whisper
ffmpeg-python
streamlit
moviepy
torch
requests
python-dotenv